+++ cron_jobs/monthlyUsersInactiver.php
@@ -1,6 +1,9 @@
 <?php
 
-// // API
+// prevent bug #95
+$_SERVER['HTTP_HOST'] = $argv[1];
+
+// API
 require_once __DIR__ . "/../vendor/autoload.php";
 require_once __DIR__ . "/../bootstrap.php";